# Optimize sqlite `StatementIterator`