--- a/diesel/src/sqlite/connection/statement_iterator.rs
+++ b/diesel/src/sqlite/connection/statement_iterator.rs
@@ -12,16 +12,74 @@ pub struct StatementIterator<'stmt, 'query> {
     field_count: usize,
 }
 
+impl<'stmt, 'query> StatementIterator<'stmt, 'query> {
+    #[cold]
+    fn handle_duplicated_row_case(
+        outer_last_row: &mut Rc<RefCell<PrivateSqliteRow<'stmt, 'query>>>,
+        column_names: &mut Option<Rc<[Option<String>]>>,
+        field_count: usize,
+    ) -> Option<QueryResult<SqliteRow<'stmt, 'query>>> {
+        // We don't own the statement. There is another existing reference, likly because
+        // a user stored the row in some long time container before calling next another time
+        // In this case we copy out the current values into a temporary store and advance
+        // the statement iterator internally afterwards
+        let mut last_row = {
+            let mut last_row = match outer_last_row.try_borrow_mut() {
+                Ok(o) => o,
+                Err(_e) => {
+                    return Some(Err(crate::result::Error::DeserializationError(
+                                    "Failed to reborrow row. Try to release any `SqliteField` or `SqliteValue` \
+                                     that exists at this point"
+                                        .into(),
+                                )));
+                }
+            };
+            let last_row = &mut *last_row;
+            let duplicated = last_row.duplicate(column_names);
+            std::mem::replace(last_row, duplicated)
+        };
+        if let PrivateSqliteRow::Direct(ref mut stmt) = last_row {
+            let res = unsafe {
+                // This is actually safe here as we've already
+                // performed one step. For the first step we would have
+                // used `PrivateStatementIterator::NotStarted` where we don't
+                // have access to `PrivateSqliteRow` at all
+                stmt.step(false)
+            };
+            match res {
+                Err(e) => Some(Err(e)),
+                Ok(false) => None,
+                Ok(true) => {
+                    let field_count = field_count;
+                    Some(Ok(SqliteRow {
+                        inner: Rc::clone(&outer_last_row),
+                        field_count,
+                    }))
+                }
+            }
+        } else {
+            // any other state than `PrivateSqliteRow::Direct` is invalid here
+            // and should not happen. If this ever happens this is a logic error
+            // in the code above
+            unreachable!(
+                "You've reached an impossible internal state. \
+                             If you ever see this error message please open \
+                             an issue at https://github.com/diesel-rs/diesel \
+                             providing example code how to trigger this error."
+            )
+        }
+    }
+}
+
 enum PrivateStatementIterator<'stmt, 'query> {
-    NotStarted(StatementUse<'stmt, 'query>),
+    NotStarted(Option<StatementUse<'stmt, 'query>>),
     Started(Rc<RefCell<PrivateSqliteRow<'stmt, 'query>>>),
-    TemporaryEmpty,
 }
 
 impl<'stmt, 'query> StatementIterator<'stmt, 'query> {
     pub fn new(stmt: StatementUse<'stmt, 'query>) -> StatementIterator<'stmt, 'query> {
         Self {
-            inner: PrivateStatementIterator::NotStarted(stmt),
+            inner: PrivateStatementIterator::NotStarted(Some(stmt)),
             column_names: None,
             field_count: 0,
         }
@@ -32,42 +90,54 @@ impl<'stmt, 'query> Iterator for StatementIterator<'stmt, 'query> {
     type Item = QueryResult<SqliteRow<'stmt, 'query>>;
 
     fn next(&mut self) -> Option<Self::Item> {
-        use PrivateStatementIterator::{NotStarted, Started, TemporaryEmpty};
-
-        match std::mem::replace(&mut self.inner, TemporaryEmpty) {
-            NotStarted(stmt) => match stmt.step() {
-                Err(e) => Some(Err(e)),
-                Ok(None) => None,
-                Ok(Some(stmt)) => {
-                    let field_count = stmt.column_count() as usize;
-                    self.field_count = field_count;
-                    let inner = Rc::new(RefCell::new(PrivateSqliteRow::Direct(stmt)));
-                    self.inner = Started(inner.clone());
-                    Some(Ok(SqliteRow { inner, field_count }))
+        use PrivateStatementIterator::{NotStarted, Started};
+        match &mut self.inner {
+            NotStarted(ref mut stmt) if stmt.is_some() => {
+                let mut stmt = stmt
+                    .take()
+                    .expect("It must be there because we checked that above");
+                let step = unsafe {
+                    // This is safe as we pass `first_step = true` to reset the cached column names
+                    stmt.step(true)
+                };
+                match step {
+                    Err(e) => Some(Err(e)),
+                    Ok(false) => None,
+                    Ok(true) => {
+                        let field_count = stmt.column_count() as usize;
+                        self.field_count = field_count;
+                        let inner = Rc::new(RefCell::new(PrivateSqliteRow::Direct(stmt)));
+                        self.inner = Started(inner.clone());
+                        Some(Ok(SqliteRow { inner, field_count }))
+                    }
                 }
-            },
-            Started(mut last_row) => {
+            }
+            Started(ref mut last_row) => {
                 // There was already at least one iteration step
                 // We check here if the caller already released the row value or not
                 // by checking if our Rc owns the data or not
-                if let Some(last_row_ref) = Rc::get_mut(&mut last_row) {
+                if let Some(last_row_ref) = Rc::get_mut(last_row) {
                     // We own the statement, there is no other reference here.
                     // This means we don't need to copy out values from the sqlite provided
                     // datastructures for now
                     // We don't need to use the runtime borrowing system of the RefCell here
                     // as we have a mutable reference, so all of this below is checked at compile time
-                    if let PrivateSqliteRow::Direct(stmt) =
-                        std::mem::replace(last_row_ref.get_mut(), PrivateSqliteRow::TemporaryEmpty)
-                    {
-                        match stmt.step() {
+                    if let PrivateSqliteRow::Direct(ref mut stmt) = last_row_ref.get_mut() {
+                        let step = unsafe {
+                            // This is actually safe here as we've already
+                            // performed one step. For the first step we would have
+                            // used `PrivateStatementIterator::NotStarted` where we don't
+                            // have access to `PrivateSqliteRow` at all
+
+                            stmt.step(false)
+                        };
+                        match step {
                             Err(e) => Some(Err(e)),
-                            Ok(None) => None,
-                            Ok(Some(stmt)) => {
+                            Ok(false) => None,
+                            Ok(true) => {
                                 let field_count = self.field_count;
-                                (*last_row_ref.get_mut()) = PrivateSqliteRow::Direct(stmt);
-                                self.inner = Started(last_row.clone());
                                 Some(Ok(SqliteRow {
-                                    inner: last_row,
+                                    inner: Rc::clone(last_row),
                                     field_count,
                                 }))
                             }
@@ -84,55 +154,19 @@ impl<'stmt, 'query> Iterator for StatementIterator<'stmt, 'query> {
                         )
                     }
                 } else {
-                    // We don't own the statement. There is another existing reference, likly because
-                    // a user stored the row in some long time container before calling next another time
-                    // In this case we copy out the current values into a temporary store and advance
-                    // the statement iterator internally afterwards
-                    let last_row = {
-                        let mut last_row = match last_row.try_borrow_mut() {
-                            Ok(o) => o,
-                            Err(_e) => {
-                                self.inner = Started(last_row.clone());
-                                return Some(Err(crate::result::Error::DeserializationError(
-                                    "Failed to reborrow row. Try to release any `SqliteField` or `SqliteValue` \
-                                     that exists at this point"
-                                        .into(),
-                                )));
-                            }
-                        };
-                        let last_row = &mut *last_row;
-                        let duplicated = last_row.duplicate(&mut self.column_names);
-                        std::mem::replace(last_row, duplicated)
-                    };
-                    if let PrivateSqliteRow::Direct(stmt) = last_row {
-                        match stmt.step() {
-                            Err(e) => Some(Err(e)),
-                            Ok(None) => None,
-                            Ok(Some(stmt)) => {
-                                let field_count = self.field_count;
-                                let last_row =
-                                    Rc::new(RefCell::new(PrivateSqliteRow::Direct(stmt)));
-                                self.inner = Started(last_row.clone());
-                                Some(Ok(SqliteRow {
-                                    inner: last_row,
-                                    field_count,
-                                }))
-                            }
-                        }
-                    } else {
-                        // any other state than `PrivateSqliteRow::Direct` is invalid here
-                        // and should not happen. If this ever happens this is a logic error
-                        // in the code above
-                        unreachable!(
-                            "You've reached an impossible internal state. \
+                    Self::handle_duplicated_row_case(
+                        last_row,
+                        &mut self.column_names,
+                        self.field_count,
+                    )
+                }
+            }
+            NotStarted(_) => unreachable!(
+                "You've reached an impossible internal state. \
                              If you ever see this error message please open \
                              an issue at https://github.com/diesel-rs/diesel \
                              providing example code how to trigger this error."
-                        )
-                    }
-                }
-            }
-            TemporaryEmpty => None,
+            ),
         }
     }
 }