This change remove a temporary state from the sqlite StatementIterator
which showed up in some profiles for our benchmarks. The benchmarks show
a 1-6% improvement in construction count for the query benchmarks and
no changes for the insert benchmarks.


<details>
<summary> Relevant benchmark results </summary>

```
bench_trivial_query/diesel/1
                        time:   [6541.7627 cycles 6542.1171 cycles 6542.5357 cycles]
                        change: [-3.0210% -2.8542% -2.7324%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 13 outliers among 100 measurements (13.00%)
  6 (6.00%) high mild
  7 (7.00%) high severe
bench_trivial_query/diesel_boxed/1
                        time:   [13976.9443 cycles 13981.9545 cycles 13986.5513 cycles]
                        change: [-1.4962% -1.4095% -1.3296%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe
bench_trivial_query/diesel_queryable_by_name/1
                        time:   [40414.1384 cycles 40456.7169 cycles 40500.9642 cycles]
                        change: [-0.7671% -0.6327% -0.5016%] (p = 0.00 < 0.05)
                        Change within noise threshold.
bench_trivial_query/diesel/10000
                        time:   [23882021.3000 cycles 23882214.7229 cycles 23882406.7944 cycles]
                        change: [-6.0633% -6.0461% -6.0331%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) low mild
  2 (2.00%) high mild
bench_trivial_query/diesel_boxed/10000
                        time:   [23890404.5122 cycles 23890792.3608 cycles 23891254.8076 cycles]
                        change: [-6.0062% -6.0030% -5.9998%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 3 outliers among 100 measurements (3.00%)
  2 (2.00%) high mild
  1 (1.00%) high severe
bench_trivial_query/diesel_queryable_by_name/10000
                        time:   [27181116.5123 cycles 27189714.1840 cycles 27198788.2537 cycles]
                        change: [-5.4711% -5.4298% -5.3904%] (p = 0.00 < 0.05)
                        Performance has improved.

bench_medium_complex_query/diesel/1
                        time:   [22495.0324 cycles 22496.0806 cycles 22497.3718 cycles]
                        change: [-1.5580% -1.4632% -1.3608%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  4 (4.00%) high mild
  7 (7.00%) high severe
bench_medium_complex_query/diesel_boxed/1
                        time:   [41632.5260 cycles 41681.6865 cycles 41730.2439 cycles]
                        change: [-1.0619% -0.9198% -0.7813%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild
bench_medium_complex_query/diesel_queryable_by_name/1
                        time:   [126045.5130 cycles 126164.1230 cycles 126292.1045 cycles]
                        change: [-0.4169% -0.2920% -0.1611%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 2 outliers among 100 measurements (2.00%)
  2 (2.00%) high mild
bench_medium_complex_query/diesel/10000
                        time:   [23650424.9443 cycles 23653430.5380 cycles 23656376.8618 cycles]
                        change: [-3.0016% -2.9868% -2.9741%] (p = 0.00 < 0.05)
                        Performance has improved.
bench_medium_complex_query/diesel_boxed/10000
                        time:   [23666598.7259 cycles 23673235.8710 cycles 23683469.4276 cycles]
                        change: [-2.9908% -2.9637% -2.9209%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 4 outliers among 100 measurements (4.00%)
  2 (2.00%) low mild
  1 (1.00%) high mild
  1 (1.00%) high severe
bench_medium_complex_query/diesel_queryable_by_name/10000
                        time:   [50436656.7219 cycles 50448976.0640 cycles 50461415.4847 cycles]
                        change: [-3.1256% -3.0928% -3.0606%] (p = 0.00 < 0.05)
                        Performance has improved.

Benchmarking bench_loading_associations_sequentially/diesel/bench_loading_associations_sequentially: Warming up for 3.0000 s
Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 5.2s, enable flat sampling, or reduce sample count to 60.
Benchmarking bench_loading_associations_sequentially/diesel/bench_loading_associations_sequentially: Collecting 100 samples in estimated 5.1724 s (5050 iter                                                                                                                                                            bench_loading_associations_sequentially/diesel/bench_loading_associations_sequentially
                        time:   [4011704.3305 cycles 4012211.9494 cycles 4012708.0785 cycles]
                        change: [-3.6397% -3.0779% -1.9779%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 15 outliers among 100 measurements (15.00%)
  4 (4.00%) low severe
  4 (4.00%) low mild
  5 (5.00%) high mild
  2 (2.00%) high severe

<details>